# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "2.0.0",
        "title": "Polanco.dev API",
        "description": "API RESTful que alimenta Polanco.dev — blog técnico, portafolio profesional, cheatsheets y trucos rápidos. Construida con Node.js, Express y MongoDB."
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1",
            "description": "Servidor de desarrollo local"
        },
        {
            "url": "https://your-app.up.railway.app/api/v1",
            "description": "Servidor de producción (Railway)"
        }
    ],
...
    "paths": {
...
        "/users": {
            "get": {
                "tags": ["Users"],
                "summary": "Listar todos los usuarios",
                "description": "Devuelve todos los usuarios registrados. Solo admin.",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Ok" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" }
                }
            }
        },
...
        "/articles": {
            "get": {
                "tags": ["Articles"],
                "summary": "Listar artículos",
                "description": "Devuelve todos los artículos publicados. Soporta filtrado, ordenamiento, paginación y selección de campos.",
                "parameters": [
                    { "name": "category", "in": "query", "schema": { "type": "string", "enum": ["Programacion", "Idioma"] }, "description": "Filtrar por categoría" },
                    { "name": "sort", "in": "query", "schema": { "type": "string" }, "description": "Ordenar por campo(s), ej: -createdAt,title" },
                    { "name": "fields", "in": "query", "schema": { "type": "string" }, "description": "Seleccionar campos, ej: title,description" },
                    { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
                    { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
                    { "name": "tags", "in": "query", "schema": { "type": "string" }, "description": "Filtrar por tags, ej: Angular,TypeScript" }
                ],
                "responses": {
                    "200": { "description": "Ok", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Article" } } } } }
                }
            },
            "post": {
                "tags": ["Articles"],
                "summary": "Crear artículo",
                "description": "Crea un nuevo artículo. Solo admin.",
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/ArticleInput" }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Artículo creado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Article" } } } },
                    "400": { "$ref": "#/components/responses/400" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" }
                }
            }
        },
...
        "/articles/{id}": {
            "get": {
                "tags": ["Articles"],
                "summary": "Obtener artículo por slug",
                "description": "Busca el artículo por su slug único, generado automáticamente desde el título al crear o actualizar el artículo.",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Slug del artículo (generado desde el título)", "example": "guia-completa-de-typescript" }
                ],
                "responses": {
                    "200": { "description": "Ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Article" } } } },
                    "404": { "$ref": "#/components/responses/404" }
                }
            },
            "patch": {
                "tags": ["Articles"],
                "summary": "Actualizar artículo",
                "description": "Actualiza parcialmente un artículo. Solo admin.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/ArticleInput" }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Artículo actualizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Article" } } } },
                    "400": { "$ref": "#/components/responses/400" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" },
                    "404": { "$ref": "#/components/responses/404" }
                }
            },
            "delete": {
                "tags": ["Articles"],
                "summary": "Eliminar artículo",
                "description": "Elimina un artículo por ID. Solo admin.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": {
                    "204": { "description": "Artículo eliminado" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" },
                    "404": { "$ref": "#/components/responses/404" }
                }
            }
        },
...
        "/articles/{articleId}/comments": {
            "get": {
                "tags": ["Comments"],
                "summary": "Listar comentarios aprobados",
                "description": "Devuelve todos los comentarios aprobados de un artículo.",
                "parameters": [
                    { "name": "articleId", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": {
                    "200": { "description": "Ok", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Comment" } } } } }
                }
            },
            "post": {
                "tags": ["Comments"],
                "summary": "Enviar comentario",
                "description": "Crea un comentario para un artículo. Queda en estado pendiente hasta ser aprobado. Requiere autenticación con rol admin.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "name": "articleId", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/CommentInput" }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Comentario enviado, pendiente de aprobación" },
                    "400": { "$ref": "#/components/responses/400" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" }
                }
            }
        },
...
        "/experiences": {
            "get": {
                "tags": ["Experiences"],
                "summary": "Obtener perfil y experiencias laborales",
                "description": "Devuelve el perfil bilingüe del CV y el listado de experiencias, ambos con campos EN/ES sin resolver.",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": { "type": "string", "example": "success" },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "profile": { "$ref": "#/components/schemas/ResumeProfile" },
                                                "experiences": { "type": "array", "items": { "$ref": "#/components/schemas/Experience" } }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Experiences"],
                "summary": "Crear experiencia",
                "description": "Solo admin. Añade una experiencia al documento singleton de Resume.",
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/ExperienceInput" }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Experiencia creada", "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "string" }, "data": { "type": "object", "properties": { "experience": { "$ref": "#/components/schemas/Experience" } } } } } } } },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" }
                }
            }
        },
...
        "/cheatsheets": {
            "get": {
                "tags": ["Cheatsheets"],
                "summary": "Listar cheatsheets",
                "parameters": [
                    { "name": "language", "in": "query", "schema": { "type": "string" }, "description": "Filtrar por lenguaje" },
                    { "name": "sort", "in": "query", "schema": { "type": "string" } },
                    { "name": "page", "in": "query", "schema": { "type": "integer" } },
                    { "name": "limit", "in": "query", "schema": { "type": "integer" } }
                ],
                "responses": {
                    "200": { "description": "Ok", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Cheatsheet" } } } } }
                }
            },
            "post": {
                "tags": ["Cheatsheets"],
                "summary": "Crear cheatsheet",
                "description": "Solo admin.",
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/Cheatsheet" }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Cheatsheet creado" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" }
                }
            }
        },
...
        "/quicktips": {
            "get": {
                "tags": ["QuickTips"],
                "summary": "Listar trucos rápidos",
                "parameters": [
                    { "name": "language", "in": "query", "schema": { "type": "string" }, "description": "Filtrar por lenguaje" },
                    { "name": "difficulty", "in": "query", "schema": { "type": "string", "enum": ["beginner", "intermediate", "advanced"] } },
                    { "name": "sort", "in": "query", "schema": { "type": "string" } },
                    { "name": "page", "in": "query", "schema": { "type": "integer" } },
                    { "name": "limit", "in": "query", "schema": { "type": "integer" } }
                ],
                "responses": {
                    "200": { "description": "Ok", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/QuickTip" } } } } }
                }
            },
            "post": {
                "tags": ["QuickTips"],
                "summary": "Crear truco rápido",
                "description": "Solo admin.",
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/QuickTip" }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "QuickTip creado" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" }
                }
            }
        },
...
        "/contact": {
            "post": {
                "tags": ["Contact"],
                "summary": "Enviar mensaje de contacto",
                "description": "Envía un email al administrador con el mensaje del formulario de contacto.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/ContactInput" }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Mensaje enviado correctamente" },
                    "400": { "$ref": "#/components/responses/400" }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "email": { "type": "string", "format": "email" },
                    "photo": { "type": "string" },
                    "role": { "type": "string", "enum": ["user", "admin"] },
                    "active": { "type": "boolean" },
                    "passwordChangedAt": { "type": "string", "format": "date-time" }
                }
            },
...
            "Article": {
                "type": "object",
                "properties": {
                    "_id": { "type": "string", "example": "615930d4f37971021d64b413" },
                    "slug": { "type": "string", "description": "Slug único generado automáticamente desde el título (solo minúsculas, separado por guiones)", "example": "guia-completa-de-typescript", "readOnly": true },
                    "title": { "type": "string", "example": "Guía completa de TypeScript" },
                    "description": { "type": "string" },
                    "content": { "type": "string" },
                    "imageCover": { "type": "string" },
                    "category": { "type": "string", "enum": ["Programacion", "Idioma"] },
                    "tags": { "type": "array", "items": { "type": "string" } },
                    "author": { "$ref": "#/components/schemas/User" },
                    "views": { "type": "integer", "default": 0 },
                    "likes": { "type": "integer", "default": 0 },
                    "published": { "type": "boolean", "default": true, "description": "false = borrador, true = publicado" },
                    "createdAt": { "type": "string", "format": "date-time" }
                }
            },
            "ArticleInput": {
                "required": ["title", "description", "category"],
                "properties": {
                    "title": { "type": "string", "minLength": 10, "maxLength": 40, "example": "Guía completa de TypeScript", "description": "Título único. El slug se genera automáticamente." },
                    "description": { "type": "string", "example": "Aprende TypeScript desde cero" },
                    "content": { "type": "string", "description": "Contenido en Markdown" },
                    "imageCover": { "type": "string", "description": "URL de la imagen de portada (obtenida de POST /upload/article-image)" },
                    "category": { "type": "string", "enum": ["Programacion", "Idioma"] },
                    "tags": { "type": "array", "items": { "type": "string" } },
                    "published": { "type": "boolean", "default": true, "description": "false para guardar como borrador" }
                }
            },
            "Comment": {
                "type": "object",
                "properties": {
                    "_id": { "type": "string" },
                    "author": { "type": "string" },
                    "email": { "type": "string", "format": "email" },
                    "body": { "type": "string" },
                    "status": { "type": "string", "enum": ["pending", "approved"], "default": "pending" },
                    "article": { "type": "string", "description": "ID del artículo" },
                    "createdAt": { "type": "string", "format": "date-time" }
                }
            },
            "CommentInput": {
                "required": ["author", "email", "body"],
                "properties": {
                    "author": { "type": "string", "example": "María García" },
                    "email": { "type": "string", "format": "email", "example": "maria@example.com" },
                    "body": { "type": "string", "example": "Excelente artículo, muy bien explicado." }
                }
            },
            "BilingualString": {
                "type": "object",
                "description": "Campo con valor en inglés y español",
                "properties": {
                    "en": { "type": "string" },
                    "es": { "type": "string" }
                },
                "required": ["en", "es"]
            },
            "BilingualStringArray": {
                "type": "object",
                "description": "Lista de cadenas bilingüe",
                "properties": {
                    "en": { "type": "array", "items": { "type": "string" } },
                    "es": { "type": "array", "items": { "type": "string" } }
                }
            },
            "Experience": {
                "type": "object",
                "description": "Experiencia laboral con campos de texto bilingüe (EN/ES sin resolver). Embebida dentro del documento singleton Resume.",
                "properties": {
                    "_id": { "type": "string", "description": "ObjectId del subdocumento embebido" },
                    "company": { "type": "string", "example": "Empresa S.A.", "minLength": 2, "maxLength": 100 },
                    "role": { "$ref": "#/components/schemas/BilingualString", "description": "Cargo/rol en ambos idiomas" },
                    "description": { "$ref": "#/components/schemas/BilingualString", "description": "Descripción del puesto en ambos idiomas" },
                    "achievements": { "$ref": "#/components/schemas/BilingualStringArray", "description": "Logros destacados en ambos idiomas" },
                    "startDate": { "type": "string", "format": "date", "example": "2022-01-01" },
                    "endDate": { "type": "string", "format": "date" },
                    "current": { "type": "boolean", "default": false }
                }
            },
            "ExperienceInput": {
                "type": "object",
                "description": "Body para crear o actualizar una experiencia embebida (campos bilingüe requeridos).",
                "required": ["company", "role", "startDate"],
                "properties": {
                    "company": { "type": "string", "example": "Empresa S.A.", "minLength": 2, "maxLength": 100 },
                    "role": {
                        "type": "object",
                        "required": ["en", "es"],
                        "properties": {
                            "en": { "type": "string", "example": "Full Stack Developer", "maxLength": 100 },
                            "es": { "type": "string", "example": "Desarrollador Full Stack", "maxLength": 100 }
                        }
                    },
                    "description": {
                        "type": "object",
                        "properties": {
                            "en": { "type": "string", "maxLength": 1000 },
                            "es": { "type": "string", "maxLength": 1000 }
                        }
                    },
                    "achievements": {
                        "type": "object",
                        "properties": {
                            "en": { "type": "array", "items": { "type": "string" } },
                            "es": { "type": "array", "items": { "type": "string" } }
                        }
                    },
                    "startDate": { "type": "string", "format": "date", "example": "2022-01-01" },
                    "endDate": { "type": "string", "format": "date" },
                    "current": { "type": "boolean", "default": false }
                }
            },
            "ResumeProfile": {
                "type": "object",
                "description": "Perfil del CV con campos bilingüe (retornado por GET /experiences sin resolver el idioma).",
                "properties": {
                    "name": { "type": "string", "example": "Diego Polanco" },
                    "email": { "type": "string", "format": "email", "example": "diego@polanco.dev" },
                    "website": { "type": "string", "example": "polanco.dev" },
                    "linkedin": { "type": "string", "example": "linkedin.com/in/dpolanco" },
                    "github": { "type": "string", "example": "github.com/dpolanco" },
                    "title": { "$ref": "#/components/schemas/BilingualString" },
                    "location": { "$ref": "#/components/schemas/BilingualString" },
                    "summary": { "$ref": "#/components/schemas/BilingualString" },
                    "skills": {
                        "type": "object",
                        "description": "Habilidades técnicas (sin traducción — iguales en ambos idiomas)",
                        "properties": {
                            "frontend": { "type": "array", "items": { "type": "string" }, "example": ["Angular", "TypeScript", "Tailwind CSS"] },
                            "backend": { "type": "array", "items": { "type": "string" }, "example": ["Node.js", "Express", "MongoDB"] },
                            "tools": { "type": "array", "items": { "type": "string" }, "example": ["Git", "Docker", "AWS"] }
                        }
                    },
                    "education": { "type": "array", "items": { "$ref": "#/components/schemas/Education" } },
                    "languages": { "type": "array", "items": { "$ref": "#/components/schemas/LanguageSkill" } }
                }
            },
            "Cheatsheet": {
                "type": "object",
                "properties": {
                    "_id": { "type": "string" },
                    "title": { "type": "string", "example": "Git Commands" },
                    "description": { "type": "string" },
                    "language": { "type": "string", "example": "git" },
                    "difficulty": { "type": "string", "enum": ["beginner", "intermediate", "advanced"] },
                    "content": { "type": "string", "description": "Contenido en Markdown" },
                    "createdAt": { "type": "string", "format": "date-time" }
                }
            },
            "QuickTip": {
                "type": "object",
                "properties": {
                    "_id": { "type": "string" },
                    "title": { "type": "string", "example": "Desestructuración de objetos en JS" },
                    "language": { "type": "string", "example": "javascript" },
                    "difficulty": { "type": "string", "enum": ["beginner", "intermediate", "advanced"] },
                    "codeSnippet": { "type": "string" },
                    "description": { "type": "string" },
                    "tags": { "type": "array", "items": { "type": "string" } },
                    "createdAt": { "type": "string", "format": "date-time" }
                }
            },
            "ContactInput": {
                "required": ["name", "email", "subject", "message"],
                "properties": {
                    "name": { "type": "string", "example": "Carlos López" },
                    "email": { "type": "string", "format": "email", "example": "carlos@example.com" },
                    "subject": { "type": "string", "example": "Consulta sobre colaboración" },
                    "message": { "type": "string", "minLength": 10, "example": "Hola Diego, me gustaría hablar sobre un proyecto..." }
                }
            },
...
            "Education": {
                "type": "object",
                "properties": {
                    "institution": { "type": "string", "example": "Universidad Nacional" },
                    "degree": { "type": "string", "example": "Ingeniería en Sistemas" },
                    "startDate": { "type": "string", "example": "2018" },
                    "endDate": { "type": "string", "example": "2022" }
                }
            },
            "LanguageSkill": {
                "type": "object",
                "properties": {
                    "language": { "type": "string", "example": "Spanish" },
                    "level": { "type": "string", "example": "Native" }
                }
            }
        },
        "responses": {
            "400": { "description": "Solicitud incorrecta — parámetros inválidos o faltantes." },
            "401": { "description": "No autorizado — token JWT faltante o inválido." },
            "403": { "description": "Prohibido — no tienes permisos para realizar esta acción." },
            "404": { "description": "No encontrado — el recurso solicitado no existe." }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token JWT obtenido en /users/login. Formato: Bearer <token>"
            }
        }
    }
}